+++ smartfixsamana-app/src/app/reparaciones/reparaciones.component.ts
@@ -7,6 +7,7 @@ import { Celular } from '../models/celular';
 import { ReparacionService } from '../services/reparacion.service';
 import { ClienteService } from '../services/cliente.service';
 import { CelularService } from '../services/celular.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-reparaciones',
@@ -16,14 +17,18 @@ import { CelularService } from '../services/celular.service';
   styleUrl: './reparaciones.component.css',
 })
 export class ReparacionesComponent implements OnInit {
-  reparaciones: Reparacion[] = [];
-  clientes: Cliente[] = [];
-  celulares: Celular[] = [];
+  reparaciones!: Reparacion[];
+
+  page: number = 0;
+  totalPages!: number;
+
+  clientes!: Cliente[];
+  celulares!: Celular[];
 
   selectedReparacion: Reparacion | null = null;
   newReparacion: Reparacion = new Reparacion();
   isAdding: boolean = false;
-  searchTerm: any;
+  keyword: string = '';
 
   constructor(
     private reparacionService: ReparacionService,
@@ -38,9 +43,14 @@ export class ReparacionesComponent implements OnInit {
   }
 
   loadReparaciones(): void {
-    this.reparacionService.getReparaciones().subscribe((reparaciones) => {
-      this.reparaciones = reparaciones;
+    this.reparacionService.getAllPageable(this.page).subscribe((response) => {
+      this.reparaciones = response.content;
+      this.totalPages = response.totalPages;
     });
+  }
+  changePage(page: number): void{
+    this.page = page;
+    this.loadReparaciones();
   }
 
   loadClientes(): void {
@@ -60,7 +70,6 @@ export class ReparacionesComponent implements OnInit {
   }
 
   getCelular(celular: Celular): string {
-    
     return celular ? `${celular.marca} ${celular.modelo}` : '';
   }
 
@@ -77,12 +86,17 @@ export class ReparacionesComponent implements OnInit {
           this.loadReparaciones();
           this.newReparacion = new Reparacion();
           this.isAdding = false;
+          Swal.fire({
+            title: 'Nueva reparación registrada!',
+            text: 'Reparación registrada con exito!',
+            icon: 'success',
+          });
         });
     } else if (this.selectedReparacion) {
       const updatePayload = {
         id: this.selectedReparacion.id,
-        clienteId: this.selectedReparacion.clienteId.id,
-        celularId: this.selectedReparacion.celularId.id,
+        clienteId: this.selectedReparacion.clienteId,
+        celularId: this.selectedReparacion.celularId,
         problema: this.selectedReparacion.problema,
         estado: this.selectedReparacion.estado,
         fechaIngreso: this.selectedReparacion.fechaIngreso,
@@ -95,17 +109,36 @@ export class ReparacionesComponent implements OnInit {
           next: () => {
             this.loadReparaciones();
             this.selectedReparacion = null;
-          },
-          error: (err) => {
-            console.error('Error actualizando la reparación:', err);
+            Swal.fire({
+              title: 'Actualizado',
+              text: 'Reparación actualizada con exito!',
+              icon: 'success',
+            });
           },
         });
     }
   }
 
   delete(reparacionId: number): void {
-    this.reparacionService.deleteReparacion(reparacionId).subscribe(() => {
-      this.loadReparaciones();
+    Swal.fire({
+      title: 'Seguro que quieres eliminar la reparación?',
+      text: 'La reparación será eliminada del sistema!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.reparacionService.deleteReparacion(reparacionId).subscribe(() => {
+          this.loadClientes();
+        });
+        Swal.fire({
+          title: 'Eliminado!',
+          text: 'Reparación eliminada con exito',
+          icon: 'success',
+        });
+      }
     });
   }
 
@@ -119,15 +152,28 @@ export class ReparacionesComponent implements OnInit {
     this.selectedReparacion = null;
   }
 
-  search(): void {
-    if (this.searchTerm) {
-      this.reparacionService
-        .getReparacionById(this.searchTerm)
-        .subscribe((data: Reparacion) => {
-          this.reparaciones = [data];
-        });
+  // Método para buscar reparaciones por cliente o celular
+  searchReparaciones(): void {
+    if (this.keyword.trim() !== '') {
+      this.reparacionService.getClienteByNombreApellido(this.keyword).subscribe(
+        (data) => {
+          this.reparaciones = data;
+        },
+        (error) => {
+          console.error('Error al buscar por cliente', error);
+        }
+      );
+
+      this.reparacionService.getReparacionByCelular(this.keyword).subscribe(
+        (data) => {
+          this.reparaciones = [...this.reparaciones, ...data]; // Combinar resultados
+        },
+        (error) => {
+          console.error('Error al buscar por celular', error);
+        }
+      );
     } else {
-      this.loadReparaciones();
+      this.loadReparaciones(); // Si no hay keyword, obtener todos los repuestos
     }
   }
 }
